Revert previous commit
The previous commit was made under the assumption that s3fs/goofys would
be a good way to deploy changes.
However it was rediculously slow, so now we are just syncing pregenerated pages.

Generating this "Coming soon" pages no longer works because it would overwrite pages
that the generator has no way of knowing about.